agent/stash_bot/requirements.txt
annotated-types==0.7.0
anthropic==0.47.1
anyio==4.8.0
asttokens==3.0.0
boto3==1.35.99
botocore==1.35.99
certifi==2025.1.31
coloredlogs==15.0.1
comm==0.2.2
dagger-io==0.17.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
docker==7.1.0
executing==2.1.0
fastapi==0.115.7
fire==0.7.0
google-genai==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langfuse==2.58.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.4.1
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytest==8.3.4
pyzmq==26.2.0
s3transfer==0.10.4
sentry-sdk==2.22.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
sentry-sdk==2.22.0
watchtower==3.4.0
